Added connections and message of the day
Initial commit for data transfer between client and server.
Added CommandMessageOfTheDay.cs
- contains additional methods and fields related to the motd
Added ConnectionHelper.cs
- contains fields and methods to handle connections

Connections are based on a simple format ("Key":"Value";). When the
client connects to the server it sends a request to obtain his specific
id, the specific server id and the message of the day (if it is set).
If we want the client (or server) to execute a certain action, we just
need to define this case in "Process<Client|Server>Data".

Changed the output of an error while executing a command for developers
in a mission screen with the complete exception.  Just thought this is
useful.