Added server configuration
Initial commit for the server configuration
Added ServerConfig.cs
- contains the configuration and handles the cfg file
Changed the header of the mission screen from 'Server: <sessionname>' to
a configurable string that is defined in the cfg file. Another
configurable string in the cfg file contains the suffix for the loaded
motd.txt that is used instead of the world id. Also in the head of the
file there is a note to which world the file belongs since its name
contains the world id.
At start the server will create the cfg file if it does not exist. The
file will contain the settings and a (detailed) description.
#22