Help, Logger and breaking changes
Enabling asteroid editing for release.
The .Size() is a breaking change, which I'm disabling for this upcoming
release as I won't be able to do it immediately after the SE 01.069
update.
Fix for the Logger creating a log file regardless of the Debug flag.
Quick update for help to display more detail. It will need more fine
tuning.