Session command changes.
CargoShips still does not appear to start after turning on whilst on a
dedicated server.
Or turning off, when it was on when DS was started.
I'm wondering if a timer event was initiated when the server is started,
which may be out of our control, which is why I've changed the Cargoship
command back to Experimental.