Request permissions if not received yet #71
- added a request for the permissions if the client hasn't received it
10 seconds after it sent the request. The client will request the
permissions every 10 second if it does not receive a response.
This way I try to adress the bug where the client does not receive the
permissions, but I'm not very sure what is causing the bug. However, it
is a feature that I wanted to implement anyway. It just seems like the
time is right for that.