#139 disable commands
Small enhancement to disable commands entirely with with `/perm command
setlevel <cmd> none`
Internally it sets the command level to uint.MaxValue, so it is still in
line with security settings, but it is an explicit value that will treat
the command as non-existent. It won't show up on help either.