Added multiword location #134
This is a bit of a kludge.
It's only falls back to multi word if it fails the initial regex
pattern, that will only happen with 3 or more words.
if you have a 2 word gps name, the first pattern will hold it.
Support.FindEntitiesNamed will need to be reworked to pass message text
back to CommandTeleport, and allow CommandTeleport to try all
combinations of teleport locations before failing.

In reality though, if you have an object called "red" another called
"one", and a third called "red one", then best of luck because we can't
tell what you mean unless we further expand the instructions for `/tp
source=red dest=one` or `/tp dest=red one` It does kind of break the
flow when you can already just type `/tp "red one"`