Ares corrections
Fixed player would only work without quotes, making it confusing how it
works. Should work either way now.
Fixed /eject not working on server. It only partially works when player
is control of Turret, so it needs a workqueue implemented.
Added /slap. It is not quite finished, as the physics push doesn't work
on the server, and even then only in 0G, not under gravity.