#156 namespace changes
This should be all namespace changes.
It appears there is an intent to phase out the usage of the Ingame
namespace from within Mod Script, so I've addressed it where ever the
Ingame is used, as a using in the namespace lists to directly make that
class available.
I have also cleaned out all old unused namespaces to prevent further
conflicts later.