Server side commands #161 #123 #120 #15
Converted many of the commands to server side, and implemented new
messages as cross channels between server and client to keep the static
details of ship lists,  player lists and /id tagged entities on the
server for easier access from now on.
Command /bomb now has the countdown timer on, to prevent the grid
floating forever.
Command /switch now has grinder and welder as optional parameters to
turn off.
Message handing for changes to a grid is handled by SyncGridChange now
to consolidate much of it.
Updated namespacing for ForceKick to keep it inline with current
codebase.
Many of the commands that previously did not work correctly since 1.105
will now work as intended. There has been extensive testing for the core
changes.
/scaleup and /scaledown have not been tested yet.
Some functionality needs to be rearranged, like /forcekick and
/forceban, which are server side now but still send a message to the
server to execute functionality. There may be others.
This update unfortunately will make it much harder to determine the flow
of logic.  Always check the command's constructor to determine if it is
Client or Server executed.